+++ src/StateSmith/compiler/EntryPointValidator.cs
@@ -4,7 +4,22 @@ public class EntryPointValidator
 {
     public static void Validate(EntryPoint state)
     {
-        PseudoStateValidator.Validate(state);
+        PseudoStateValidator.ValidateParentAndNoChildren(state);
         PseudoStateValidator.ValdiateEnteringBehaviors(state);
+    }
+}
+
+public class ExitPointValidator
+{
+    public static void Validate(ExitPoint v)
+    {
+        PseudoStateValidator.ValidateParentAndNoChildren(v);
+
+        if (v.IncomingTransitions.Count == 0)
+        {
+            throw new VertexValidationException(v, "An exit point must at least one incoming transition (for now).");
+        }
+
+        PseudoStateValidator.ValdiateExitingBehaviors(v);
     }
 }