+++ src/StateSmithTest/EntryPointValidationTests.cs
@@ -1,8 +1,10 @@
 ﻿using StateSmith.Compiling;
+using System.Linq;
+using Xunit;
 
-namespace StateSmithTest.InitialStateProcessor
+namespace StateSmithTest.PseudoStateTests
 {
-    public class EntryPointValidationTests : PseudoStateValidationTestHelper
+    public class EntryPointValidationTests : EntryInitialValidationTestHelper
     {
         public EntryPointValidationTests()
         {
@@ -13,6 +15,24 @@ namespace StateSmithTest.InitialStateProcessor
         {
             return new EntryPoint("1");
         }
+    }
 
+    public class ExitPointValidationTests : PseudoStateValidationTestHelper
+    {
+        public ExitPointValidationTests()
+        {
+            s2_1.AddTransitionTo(s2_pseudoState);
+            AddBlankS2PseudoStateTransition(); // to make exit point validation successfull
+        }
+
+        protected override void AddBlankS2PseudoStateTransition()
+        {
+            s2_pseudoState.AddTransitionTo(s1);
+        }
+
+        override protected ExitPoint CreateS2PseudoState()
+        {
+            return new ExitPoint("1");
+        }
     }
 }