+++ src/StateSmithTest/spec/2/c/main.c
@@ -22,9 +22,10 @@ const char * const event_names[] = {
     [Spec2Sm_EventId_EV6] = "EV6",
     [Spec2Sm_EventId_EV7] = "EV7",
     [Spec2Sm_EventId_EV8] = "EV8",
+    [Spec2Sm_EventId_EV9] = "EV9",
 };
 static_assert(COUNT_OF(event_names) == Spec2Sm_EventIdCount, "above mapping outdated");
-static_assert(Spec2Sm_EventIdCount == 9, "above mapping outdated");
+static_assert(Spec2Sm_EventIdCount == 10, "above mapping outdated");
 
 
 int main(int arg_count, char** args)